+++ app/Domain/Experience/Reactors/ExperienceAchievementReactor.php
@@ -2,7 +2,7 @@
 
 namespace App\Domain\Experience\Reactors;
 
-use App\Domain\Achievements\Experience\ExperienceAchievementUnlocker;
+use App\Domain\Experience\Models\Achievement;
 use App\Domain\Experience\Commands\UnlockAchievement;
 use App\Domain\Experience\EventQueries\ExperienceAmountQuery;
 use App\Domain\Experience\Events\ExperienceEarned;
@@ -13,7 +13,6 @@ class ExperienceAchievementReactor extends Reactor
 {
     public function __construct(
         protected CommandBus $bus,
-        protected ExperienceAchievementUnlocker $unlocker
     ) {
     }
 
@@ -21,7 +20,7 @@ class ExperienceAchievementReactor extends Reactor
     {
         $query = new ExperienceAmountQuery($event->aggregateRootUuid());
 
-        $achievement = $this->unlocker->achievementToBeUnlocked(
+        $achievement = $this->resolveAchievement(
             previousCount: $query->previousCount(),
             currentCount: $query->currentCount(),
             userId: $event->userId,
@@ -36,5 +35,18 @@ class ExperienceAchievementReactor extends Reactor
             userId: $event->userId,
             achievement: $achievement,
         ));
+    }
+
+    protected function resolveAchievement(
+        int $previousCount,
+        int $currentCount,
+        int $userId
+    ): ?Achievement {
+        return Achievement::forExperience()
+            ->get()
+            ->filter(fn (Achievement $achievement) => $previousCount < $achievement->data['count_requirement'])
+            ->filter(fn (Achievement $achievement) => $currentCount >= $achievement->data['count_requirement'])
+            ->reject(fn (Achievement $achievement) => $achievement->receivedBy($userId))
+            ->first();
     }
 }
