+++ app/Domain/Experience/Reactors/SeriesAchievementReactor.php
@@ -2,12 +2,11 @@
 
 namespace App\Domain\Experience\Reactors;
 
-use App\Domain\Achievements\Series\SeriesCompletionAchievementUnlocker;
+use App\Domain\Experience\Models\Achievement;
 use App\Domain\Experience\Commands\UnlockAchievement;
-use App\Domain\Experience\EventQueries\ExperienceAmountQuery;
-use App\Domain\Experience\Events\ExperienceEarned;
 use App\Domain\Experience\Events\SeriesCompleted;
 use App\Models\Series;
+use App\Models\User;
 use Spatie\EventSourcing\Commands\CommandBus;
 use Spatie\EventSourcing\EventHandlers\Reactors\Reactor;
 
@@ -15,7 +14,6 @@ class SeriesAchievementReactor extends Reactor
 {
     public function __construct(
         protected CommandBus $bus,
-        protected SeriesCompletionAchievementUnlocker $unlocker
     ) {
     }
 
@@ -23,7 +21,7 @@ class SeriesAchievementReactor extends Reactor
     {
         $series = Series::find($event->seriesId);
 
-        $achievement = $this->unlocker->achievementToBeUnlocked(
+        $achievement = $this->resolveAchievement(
             series: $series,
             userId: $event->userId,
         );
@@ -37,5 +35,28 @@ class SeriesAchievementReactor extends Reactor
             userId: $event->userId,
             achievement: $achievement,
         ));
+    }
+
+    protected function resolveAchievement(
+        Series $series,
+        int $userId,
+    ): ?Achievement {
+        $achievement = Achievement::forSeries($series)->first();
+
+        if (! $achievement) {
+            return null;
+        }
+
+        if ($achievement->receivedBy($userId)) {
+            return null;
+        }
+
+        $user = User::find($userId);
+
+        if (! $user->hasCompleted($series)) {
+            return null;
+        }
+
+        return $achievement;
     }
 }
